Split the math module, removed unnecessary include paths and refactored a few bazel targets that were too specific.
