+++ .travis.yml
@@ -21,4 +21,4 @@ before_install:
 
 script:
   - bazel build --verbose_failures ...
-  - bazel test ...
+  - bazel test --test_verbose_timeout_warnings ...
