+++ common/BUILD.bazel
@@ -3,7 +3,6 @@ cc_library(
     hdrs = [
         "include/types.h",
     ],
-    copts = ["-Icommon/include"],
     visibility = ["//visibility:public"],
 )
 
@@ -12,7 +11,6 @@ cc_library(
     hdrs = [
         "include/error_macros.h",
     ],
-    copts = ["-Icommon/include"],
     visibility = ["//visibility:public"],
 )
