+++ math/BUILD.bazel
@@ -1,9 +1,8 @@
 cc_library(
     name = "matrix_math",
     hdrs = [
-        "include/matrix_checks.h"
+        "matrix/include/matrix_checks.h"
     ],
-    copts = ["-Imath/include"],
     deps = [
         "//common:types",
     ],
@@ -11,11 +10,10 @@ cc_library(
 )
 
 cc_library(
-    name = "incremental_average",
+    name = "misc_math",
     hdrs = [
-        "include/incremental_average.h"
+        "misc/include/incremental_average.h"
     ],
-    copts = ["-Imath/include"],
     deps = [
         "//common:macros",
         "//common:types",
