+++ math/misc/include/incremental_average.h
